# specs/003-add-a-matching/contracts/api.yaml
openapi: 3.0.3
info:
  title: Admin Dashboard API
  version: 1.0.0
  description: |
    API contracts for the admin dashboard content management system.
    Extends existing blog and portfolio APIs with admin-specific operations.
  contact:
    name: Rohan Batra
    email: contact@rohanbatra.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://rohanbatra.com/api
    description: Production server
security:
  - ClerkAuth: []
paths:
...
  /admin/activity:
    get:
      summary: Get recent system activity
      tags: [Admin Dashboard]
      security:
        - ClerkAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 50
          description: Number of recent items to return per category
      responses:
        '200':
          description: Recent activity retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecentActivity'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # Blog Post Management (Admin Extensions)
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      description: Clerk JWT token in Authorization header
  schemas:
...
    RecentActivity:
      type: object
      properties:
        recentComments:
          type: array
          items:
            $ref: '#/components/schemas/AdminComment'
        recentUsers:
          type: array
          items:
            $ref: '#/components/schemas/AdminUser'
        recentPosts:
          type: array
          items:
            $ref: '#/components/schemas/AdminBlogPost'
        recentProjects:
          type: array
          items:
            $ref: '#/components/schemas/AdminProject'

    # Blog Post Admin Schemas
    AdminBlogPost:
      allOf:
        - $ref: '#/components/schemas/BlogPost'
        - type: object
          properties:
            analytics:
              type: object
              properties:
                viewCount:
                  type: integer
                likeCount:
                  type: integer
                commentCount:
                  type: integer
                lastViewed:
                  type: string
                  format: date-time
            author:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                email:
                  type: string
            moderation:
              type: object
              properties:
                flaggedComments:
                  type: integer
                pendingComments:
                  type: integer
    BlogPost:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        featuredImage:
          type: string
        images:
          type: array
          items:
            type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, published, archived]
        featured:
          type: boolean
        seoTitle:
          type: string
        seoDescription:
          type: string
        readingTime:
          type: integer
        authorId:
          type: string
        publishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AdminProject:
      allOf:
        - $ref: '#/components/schemas/Project'
        - type: object
          properties:
            analytics:
              type: object
              properties:
                viewCount:
                  type: integer
                lastViewed:
                  type: string
                  format: date-time
            author:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
    Project:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        longDescription:
          type: string
        category:
          type: string
        technologies:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, published, archived]
        featured:
          type: boolean
        images:
          type: array
          items:
            type: string
        demoUrl:
          type: string
        sourceUrl:
          type: string
        liveUrl:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        client:
          type: string
        tags:
          type: array
          items:
            type: string
        authorId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AdminComment:
      allOf:
        - $ref: '#/components/schemas/Comment'
        - type: object
          properties:
            post:
              type: object
              properties:
                slug:
                  type: string
                title:
                  type: string
            user:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                email:
                  type: string
                avatar:
                  type: string
            moderation:
              type: object
              properties:
                flagCount:
                  type: integer
                moderatedBy:
                  type: string
                moderatedAt:
                  type: string
                  format: date-time
                moderationReason:
                  type: string
    Comment:
      type: object
      properties:
        _id:
          type: string
        postSlug:
          type: string
        userId:
          type: string
        content:
          type: string
        parentId:
          type: string
        status:
          type: string
          enum: [pending, published, spam, deleted]
        language:
          type: string
        metadata:
          type: object
          properties:
            ipAddress:
              type: string
            userAgent:
              type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AdminUser:
      type: object
      properties:
        _id:
          type: string
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, editor, admin]
        avatar:
          type: string
        bio:
          type: string
        preferences:
          type: object
        activity:
          type: object
          properties:
            commentCount:
              type: integer
            likeCount:
              type: integer
            lastActive:
              type: string
              format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
...
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...